Fixed Resize and Minimum Sizing
Previously the minimum size for every plugin panel was defined by the
largest plugin, however this caused issues with resizing smaller plugins
and required all plugins to fill that space.  These changes correct this
by utilizing the "minimum size" parameter for each plugin panel.  If a
minimum size is not defined for a panel then the old method of sizing
for the largest plugin is used instead.